webapp/requirements.txt
Django==4.2.9
psycopg2-binary>=2.9.9
redis==5.0.1
django-redis==5.4.0
requests==2.31.0
beautifulsoup4==4.12.3
whitenoise==6.6.0
python-dotenv>=1.0.0
anthropic>=0.40.0
playwright>=1.40.0
